requirements.txt
python-dotenv
pyfiglet
discord.py